OMAPDSS: DSS: SyncLost WA for wb m2m-manager mode.
This is a workaround that prevents SYNC_LOST on changing pipe
channelout from manager which was used as a source of wb to
another manager. Manager could free pipes after wb will
send SYNC message but that will start wb capture. To prevent
that we reconnect the pipe from the manager to wb and do a
dummy enabling and disabling of wb - the pipe will be freed
and capture won't start because source pipe is switched off.

Change-Id: Ie70105dcf7c57f68db2e1342ff7288741dd1c109
Author: Artem Rudenko <x0150883@ti.com>
Signed-off-by: Mykhailo Denysiuk <x0172934@ti.com>
Signed-off-by: Artem Rudenko <artemrudenko@ti.com>