ARM: OMAP5430: hwmod data: Remove unneeded ADDR_TYPE_RT
ADDR_TYPE_RT is used to find the address range that will
contain the SYSCONFIG register.
In case of dual mapping, only the direct path from MPU
does need that flag.
In case the SYSCONFIG does not exist, this flag is not
needed either.

Remove it if the above conditions are met.

Signed-off-by: Benoit Cousson <b-cousson@ti.com>
[only update change-ID for review.omapzoom]
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>

Change-Id: Ib7f4daca34776ab1d2898b496280b9e8e588e7d7