rpmsg: omx: fix check in rpmsg_omx_release function for destroying ept
Checking for !omxserv->rpdev to know if the release function has to
destroy the ept is not valid since, a new rpdev is created after a
recovery. So, in case userspace closes its omx handle after the rproc
recovery has happened, a new rpdev is associated to the omxserv and
when rpmsg_omx_release is executed the check !omxserv->rpdev will be
true trying to destroy the ept again and causing a kernel panic. This
patch is replacing the check with omx->state != OMX_FAIL which is
safe.

Change-Id: Id49c25aa5bd66c01dd7a865c9ff0e1bdde910517
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>