TI thermal framework: take into account the amount of OPPs
In order to define the amount of steps we will do on the
zone above sys threshold hot, this patch changes the
code to consider the amount of opps available on cpu
and gpu. Now the amount of steps will be based on
the amount of opps + 1 extra for the thermal shutdown
due to the case policy.

Change-Id: Id3a6cb7c944c0fd7d7b51bf59073145ee8b6f148
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>