mfd: twl6030: add support for power resources
TWL6030 resources are designed to be used with multiple processors
we should ideally be able to describe the way the board is wired up.

The default configuration which is almost universally used is to
connect TWL6030 only to Application Processor(example: OMAP4430/4460)
and the request lines for 2(a,b,c) and 3 are completely unused.

This patch adds ability to manage the power resources of the
TWL6030.

Change-Id: I1a099577bca3c73d0332259b0f15b4bdd9633457
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Ambresh K <ambresh@ti.com>
Signed-off-by: Baek.Kyung-Han <wildtaz.sec@gmail.com>
[oleksandr.dmytryshyn@ti.com: ported from K3.0]
Signed-off-by: Oleksandr Dmytryshyn <oleksandr.dmytryshyn@ti.com>