omap5: bandgap: fix context restore function
In the context restore function, if the context is lost or
not is being checked by the contents of the counter register.
But this is logic hold good as long as counter reset value is
zero, if the reset value is non-zero then above logic doesn't
hold good. Hence removed checking of the register value and
restoring the context.

Change-Id: I74401199aedd71e3654eb3a2d5f54b7992517670
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>