ARM: OMAP4: PM: Handle LPDDR CLK IO errata during suspend/idle
There is a HW bug in CMD PHY which gives ISO signals as same for both
PADn and PADp on differential IO pad, because of which IO leaks higher
as pull controls are differential internally and pull value does not
match A value. Though there is no functionality impact due to this bug,
it is seen that by disabling the pulls there is a savings ~500uA in
OSWR, but draws ~300uA more during OFF mode.

To save power during both idle/suspend following approach taken:
1) Enable WA during boot-up.
2) Disable WA while attempting suspend and enable during resume.

TI EID: i736; IMPACTED: all OMAP4

Change-Id: Ie361446ed4d70ec4183f916fba880ab4049ba28f
Reported-by: Logan Lee <hklee@ti.com>
Signed-off-by: Shashi Shekar B.S. <shashi.shekar@ti.com>
Signed-off-by: Ambresh K <ambresh@ti.com>
[grygorii.strashko@ti.com - Forward port from K3.0]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>