TI thermal framework: on die domains: avoid spurious temperature reports
If the domain receives several spurious temperature reports, this may
cause policy behavior mismatch, i.e. over throttling. As of today,
every time the sensor is reporting its temperature, it will also send
the value to the corresponding governor. In turns this causes misbehavior
in cases where we would not expect the policy to react, for instance
when reading temperature from debugfs.

This patch changes the way the domain works. Now the governor is expected
to be notified only when there is a temperature alert coming from IRQ
or when the timer on the averaging working queue is expired.

Change-Id: I36264e7433020450feec7e5838c20b1b044d92e9
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>