omap5: bandgap: introduce has_tshut_config flag
This has_tshut_config flag is introduced to protect
access to tshut thresholds. On OMAP5 ES1.0 these tshut
registers need to be configured by SW but on ES2.0
the reset values are loaded from efuse and SW need not
touch these registers. Hence this patch introduces a
flag to control the read/write access to tshut thresholds.

Change-Id: Icd7bc51f2cfa463c9e8787540c2e01ff46e8d403
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>