omap5: bandgap: PPA_TSHUT mechanism due to COBRA-BUG-175
Due to HW bug in omap5 ES1.0, the tshut mechanism is implemented
using PPA. And to achieve this, the HLOS should not touch the
alert and tshut registers. This patch prevents modifying any of
the alert and tshut registers on Non-GP device.

On Non-GP device, this HLOS patch and the new PPA with TSHUT feature
SHOULD be used together, otherwise the system will not boot.

On GP devices, this patch will not have any effect and thresholds
will be under HLOS control.

Change-Id: I3a6782bdd3a097931b643b3ee1c278d66cadfe26
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>