android_omap_defconfig: Enable uHID
uHID driver adds support user-space I/O drivers for the
HID subsystem. This allows to write transport-level drivers like USB-HID
and Bluetooth-HID in user-space.

Low-Energy Bluetooth needs this to feed HID data that is parsed in
user-space back into the kernel.
Bluetooth stacks based on user-space also make use of this driver to feed HID
data that is parsed in user-space back into the kernel.

Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>
Change-Id: I68e478c422e82c213bd3eee2a4ebc70561977eb9