TEMP WA:OMAP: SAR: save USB UHH context conditionally
This patch adds support for SAR module / entry flags to the core code.
Support for SAR_SAVE_COND and SAR_INVALID flags is added at this point.
SAR_SAVE_COND is needed for fixing a USB UHH related bug, where the
save-and-restore context can be corrupted if the device enters off mode
repeatedly.

This errata exists on all OMAP4 and 5 variants.
The next version of this errata fix is expected to be in the scope of the
EHCI driver.

Portions of the patch are ported from this commit id on K3.0

(From b5edcef7f15a12302b9cecb13478955641cb8d2d Mon Sep 17 00:00:00 2001
From: Jon Hunter <jon-hunter@ti.com>
Date: Tue, 6 Mar 2012 17:30:58 -0600
Subject: [PATCH] ARM: OMAP4: USB: Update workaround for USB errata i719

Due to USB errata i719 the USB host save-and-restore context can be corrupted
if the device enters off mode more than two times without enabling the USB
host controller in between.

The corruption is occuring during the software save of the USB context which
occurs on entry to every off mode attempt. To avoid this problem only save the
USB context if the USB host controller has been enabled since the previous off
mode attempt. We do not need to save the USB context on every entry to off mode
if the USB controller has not been enabled.

Signed-off-by: Jon Hunter <jon-hunter@ti.com>
Signed-off-by: Volodymyr Riazantsev <v.riazantsev@ti.com>
Signed-off-by: Ruslan Bilovol <ruslan.bilovol@ti.com>
)

Change-Id: I7cbdc1915e3208c6f63a8137236d63fceed45111
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>
Signed-off-by: Tero Kristo <t-kristo@ti.com>