mmc: omap: Use omap hsmmc API to scale OPP and clock
Use the omap platform specific OPP scaling API to scale to
appropriate OPP as necessary. For example, OMAP5 MMC1/2
are not timing closed to use 192MHz functional clock
at OPP_50. So, we use this API and let the platform
specific functions handle if scaling to higher OPP
is necessary and take any action accordingly.

Change-Id: I9f75b6f50f77f44760767ff3ba87815512db4485
Signed-off-by: Viswanath Puttagunta <vishp@ti.com>