OMAPDSS: DISPC: Add force 1D tiled mode support for omap5 ES2.0 silicon
Descrition: omap5 es2.0 silicon allows 1d accesses to tiled regions. This is
done to allow skipping multiple lines in tiled address space.

the ddma of es1.0 silicon generates ocp bus error upon trying to do 1d access
in tiled address space. This limitation is removed in es2.0 silicon by adding
a new bit field FORCE1DTILEDMODE which allows the ddma to do 1d access in tiled
address space.

This patch adds framework support for accessing this new bit field. The actual
implementation of this feature is application specific and is not part of this
patch.

Change-Id: I528883ba8ac94508a7269841cb56f99ad53b6602
Signed-off-by: Sreenidhi Koti <sreenidhi@ti.com>