TI thermal framework: introduce temperature injection for debugging
For debugging purposes, it is interesting to have a way to exercise
the policies without waiting for the silicon to change it's thermal
dynamics. This patch introduces the temperature injection feature.

Now, it should be possible to write scripts to change the temperature
for a specific sensor, consequently of a domain and its governor is
expected to behave accordingly.

The sensor will be reporting the same value written by the user.
A negative value will return the sensor to it's normal operation.

Change-Id: I9b8eb991f714951b8f6f83a1a21d6ff5525562a7
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>