HACK: GPIO: OMAP5432: Remove error message when WA cannot be applied
This message floods the console on Panda5 boards because we fail
to apply the GPIO workaround on GPIO65.  GPIO65 is not a valid
pin on OMAP5432 since the HSI1 interface is not brought out to any
balls on the chip.  We keep this pin configured in safe mode but
ducati is configuring the gpio direction since this is a camera
pin on sEVM/5430 and we share the same code.

We need Ducati to configure the pin correctly but that is not
an option for this RC as that code base has been locked down.
So we'll workaround the issue by removing this message
until Ducati can mux the pins correctly for Panda5.

Since the WA kicks in both idle & suspend path, if the print
messages keep coming it will prevent lower idle states.
Also as mentioned before print_ratelimit will only limit the
prints but since WA path is taken in idle path too they will
keep coming.
The best way for now is to keep this hack as it is updated
here and get it fixed in Ducati

Change-Id: If3a43305822a721fc1e95dc13c35108e590bd69b
Signed-off-by: Kevin McCray <kevin.mccray@ti.com>