mfd: twl6030-gpadc: Fix GPADC channel 0 gain
Fixed the gain value for GPADC channel 0.
Affected only TWL6030.
The gain was set to the value 1142, and it was a bug.
This constant should be defined as 1.25*8=10,
where 1.25 is Vref (1.25V) and 8 is multiplicant
for fixed poind operation.
This channel is used for battery detection.
Thus the bug caused the wrong detection of the battery
instead of the emulator if working on emulator.

Change-Id: I1eb221cc35d50c3fd366dea11f9f8c756669b4b1
Signed-off-by: Oleg_Kosheliev <oleg.kosheliev@ti.com>
Signed-off-by: Balaji T K <balajitk@ti.com>