twl603x_charger: Integrate Fuel Gauge
Integrate Fuel Gauge in twl603x_bci_battery driver.

The driver measures the parameters needed for FG,
maintains a periodical calls of the FG process
functions and provides a charge status info
for proper working of the FG.
Also FG can use driver functions to stop
charging at "charge complete" condition and to
restart it again when discharged to restart level.
A cell_state data is integrated in device info
structure. The Fuel Gauge data is in the sysfs
group which is under the control of the
twl603x_bci_battery driver.
This patch completely removes the previous
implementation of fuel gauging.
If CONFIG_FUEL_GAUGE is not defined in defconfig,
or if the battery is not present
the state of charge parameter is set to 100%.

Change-Id: If399c9a278bbbf0dd043462f749ec42dadc46eab
Signed-off-by: Max Herman <mherman@fsisys.com>
Signed-off-by: Oleg_Kosheliev <oleg.kosheliev@ti.com>