ARM: OMAP5: ES1: hwmod_data: Move hwmod data files to *_es1
This patch is to duplicate hwmod data and other relative 54xx
data files and their headers.

This is a temporary patch to have both es1 and es2 hwmod data
in one baseline.

After ES2 wakeup .. this patch will be removed

and moving forward,only ES2 support will be continued on this
baseline

Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>
Change-Id: Ibd018a10ffabc522ff049da0f9a8463742756d67