ARM: OMAP5: hwmod data: Introduce ES2.0 initial version
This patch is to introduce the initial version of
OMAP54XX ES2.0 hwmod data.

ES1.0 and ES2.0 co-existence in a combined version of data
source is complex to do. As the amount of changes especially
involved with L4PER moving into CORE domain.

Hence using the default supported version of hwmod as for
OMAP5430 ES2.0

This patch is generated with

Author: Benoit Cousson <b-cousson@ti.com>
git: omap_autogen
branch: omap5430_for_3.8
location: outputs/omap5430/2.0/android

Change-Id: I8751644ca34a25f67ce5484885663c1a36d803ff
[porting to p-linux-omap-3.4 integrated mainline]
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>
[Author of the hwmod tool for ES2.0]
Signed-off-by: Benoit Cousson <b-cousson@ti.com>