usb: musb: Dont start Host session on probe unless device attached
This patch makes sure that we dont start the Host session unless
there is a cold boot scenario and a device(mouse) is attached.

Without this patch, the Host session was started by default
causing the gadget configuration to be not recognized by
the Host machine on cold boot in device mode.

Based on k3.0 commit:
p-android-omap-3.4: ddf6b5ae23 [Moiz Sonasath <m-sonasath@ti.com>]

Change-Id: I11c1c82b55424b0a0bc250e7e9c2c752854f08ca
Signed-off-by: Alexander Savchenko <oleksandr.savchenko@ti.com>