omap: bandgap: reorganize defines for register offsets
The bandgap register offsets and bit masks are already defined
in ctrl_module_core_54xx.h file. Modified the BG driver code
to use those offsets and bit masks.

Besides, this patch also prepares the code to better differentiate
between ES1.0 definitions and ES2.0 definitions. Idea is to
resolve the difference at ctrl_module_core_54xx.h. When the build
is for ES1.0, it will include the proper header. Users of this
header will not need to worry about which file to include.

This patch makes sure support for ES1.0 is kept in one piece.
Support for ES2.0 will be added later.

Change-Id: Ib8e3234dc3dd22ebb31a07d564f194b1e9a20d30
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>