thermal: omap: reorganization of data structures
This patch re-structures the needed data so the code more
readable. The data structures needed for each omap arch
type has been moved to its own data file.

Some changes in the data structures definitions are also
done for simplicity and for coping with the nature of the
data structures.

Change-Id: I4fb0f0ab037017be56cb8b8d3edc648c7e983bcc
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>