omap5: bandgap: add ES2.0 changes
The bandgap module in OMAP5 ES2.0 has changed.

Following are the major changes:
1. Clock source has changed. The clock is sourced from L3_INSTR.
2. The Counter register has changed, there is no dedicated register.
Now it is part of mask register.
3. The ADC conversion/lookup table has changed.
4. New registers are added, Cumulative register and History buffer.

This patch makes sure that the driver code works on both ES1.0 and
ES2.0 with same functionality. This patch DOES-NOT address making use
of new features introduced (like Cumulative & History Buffer) in ES2.0.

Change-Id: Ibea3e04d18eedf0d9ed86fe1d6be4e6448aad3a8
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>