MFD: twl-core: Add ability to set one-time registers configuration
On some platforms, some the TWL registers may need to be configure once
during the initialization. Each time when we need to do that - struct
twl4030_platform_data and twl-core files need to be modified, because,
not all TWL registers/parameters are defined now in struct
twl4030_platform_data.

Hence, add common mechanism to set TWL one-time registers configuration
which will be applied at the end of TWL initialization. And to define
registers configuration in platform code as:
struct twl_reg_setup_array reg_setup_script = {
  {mod_no, addr, u8 val, desc;}     <- register 1
  {mod_no1, addr1, u8 val, desc1;}  <- register 2
  {0, 0, 0, NULL} <- terminator
}
and pass it to the twl-core through twl4030_platform_data.

Change-Id: Ia1b9ffa92dfe24b65c846d99aaaa9130f1e25671
Signed-off-by: Nishanth Menon <nm@ti.com>
[oleksandr.tyshchenko@ti.com: ported from K3.0]
Signed-off-by: Oleksandr Tyshchenko <oleksandr.tyshchenko@ti.com>
[grygorii.strashko@ti.com: made it common and move in TWL core]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>