ARM: OMAP5: DPLL: Fix MPU DPLL rate check
omap5_mpu_dpll_set_rate() function skips setting
MPU DPLL rate if it is already the same.
But rate check is not correct if DCC is used.

For example we set MPU frequency to 1500MHz (DCC is used).
If we try to set 1500MHz again then configuration should be skipped.
But it is not skipped:

    rate = 1500;
    dpll_rate = 750;
    if (rate == dpll_rate/2) -> (1500 != 375)

Instead check should be changed to (rate/2 == dpll_rate)

Change-Id: I7e2115b2f1024915d9efcd35add56cc30a5feb0e
Signed-off-by: Taras Kondratiuk <taras@ti.com>