ARM: OMAP5: hwmod: add ocp2scp child device information as dev_attr
ocp2scp driver probe takes care of creating the child devices connected to it.
In omap5, ocp2scp has usb2 and usb3 phys connected to it. This information is
passed as pdata to ocp2scp driver. So filled the hwmod data with the
child device information which is passed in pdata when creating the
ocp2scp device.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>