ARM: OMAP5: hwmod data: cleanup duplicate mmu hwmods for ipu & dsp
The mmu hwmod data for ipu and dsp are defined twice using different
names, one with a 'mmu' in the name and the other without. The 'mmu'
ones are defined newly as part of the new ES2 data. All the duplicate
structures using this name are removed, and the original ones are
retained. This is done to keep the structures similar to the one in
the ES1 data file, even though the 'mmu' ones are the ones that
would show up eventually in the upstream kernel.

Change-Id: Ibad9c5735bf917c4dfb21a4b0346b46cd1cb05ba
Signed-off-by: Suman Anna <s-anna@ti.com>