ARM: OMAP54XX: PM: handle device instance offset in ES2.0 for non-EMIF I/O isolation
OMAP543x ES2.0 has a different PRM_DEVICE_INST offset
compared to ES1.0 OR OMAP4. ES1.0 Vs ES2.0 is compile defined,
However, we need to support OMAP4 Vs OMAP5 runtime, so handle
instance handling based on cpu.

The rest of the parameters are the same, so we leave them alone.

Change-Id: Ibbb99d1dbb055c11e4a598451f771e2d6b0918c2
Signed-off-by: Nishanth Menon <nm@ti.com>