ARM: OMAP5: hwmod data: remove omap54xx_ipu_sysc
The omap54xx_ipu_sysc structure is defined and used with the
omap54xx_ipu_hwmod_class. This class is used with the ipu
processor pseudo-hwmods, which are defined only for reset
purposes. The sysc structure need not be defined for these
hwmods and is therefore cleaned up. An equivalent sysc structure
is defined and used with the main ipu hwmod that also controls
the ipu mmu.

Change-Id: I44d79c18e98efa2d398675dfc8ef91044a4c820a
Signed-off-by: Suman Anna <s-anna@ti.com>