ARM: HACK: OMAP54XX: opp data: update rate for iva clock
Update DPLL_IVA rate for OPP_NOM. This is needed to make
round rate function working. For now round rate function
tries to choose the rate value which is very close to defined
one, but not the same, and it fails. Patch updates the rate
to make it working. Should be removed after proper rootcause
and fix of round rate function.

2330000000 -> 2329600000.

Change-Id: I16d02a2835b2a14fb70e80851b2fcac65bcffc33
Signed-off-by: Andrii.Tseglytskyi <andrii.tseglytskyi@ti.com>