ARM: OMAP4: twl-common: Configure some resources as always_on
The regulator framework will shut down any unused regulators
in the system. This was causing the some regulators to be shut
down, but VDAC supplies clock generation chip, VANA supplies
GPADC and GAS-GAUGE, CLK32KG is used for the WILINK, etc.
So configure resources VDAC, VANA, CLK32KG and V2V1 as always_on
by default since it was implemented in K3.0.

Change-Id: Ib3ca353fb91033c6b68e01f137d9ea31b49ff5d1
[balajitk@ti.com: add always_on for VANA]
Signed-off-by: Balaji T K <balajitk@ti.com>
[m-sonasath@ti.com: add always_on for VDAC]
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>
[vishalm@ti.com: add always_on for CLK32KG]
Signed-off-by: Vishal Mahaveer <vishalm@ti.com>
[oleksandr.tyshchenko@ti.com: ported from K3.0]
Signed-off-by: Oleksandr Tyshchenko <oleksandr.tyshchenko@ti.com>