ARM: OMAP4: twl603x: Enable SMPS standard mode with offset by default
In TWL603x depending on the value of SMPS_OFFSET efuse register the SMPS
regulator voltage range supported in standard mode can be either between
~0.6V - ~1.3V or ~0.7V - ~1.4V. In TWL6030 ES1.0, SMPS_OFFSET efuse is
programmed to all 0's where as starting from TWL6030 ES1.1 the efuse is
programmed to 1.

The wider voltage range should be enabled on OMAP4 for supporting of
higher OPPs, hence always try to enable standard mode with offset for
SMPS regulators. Because twl603x_set_offset() need to be called only once -
run it for MPU VD on OMAP4430/70 and Core VD on OMAP4460
(where MPU VD controlled by TPS6236X).

In case we are unable to do anything about the offset, then we throw
out a warning and continue as the system should still be able to
function despite the higher value configuration.

Change-Id: I710ca4e4b7b83701465c2aab322803a6bd2476e3
[nm@ti.com: rewrite, split up and new controls]
Signed-off-by: Nishanth Menon <nm@ti.com>
[oleksandr.kozaruk@ti.com: original patch]
Signed-off-by: Oleksandr Kozaruk <oleksandr.kozaruk@ti.com>
[oleksandr.tyshchenko@ti.com: ported from K3.0]
Signed-off-by: Oleksandr Tyshchenko <oleksandr.tyshchenko@ti.com>
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>