ARM: OMAP5: sevm: Fix usage slope/offset values for tmp102
As there was the issue in tmp102 driver with wrong usage of
slope_cpu/offset_cpu for domains other then CPU and slope/offset
for CPU domain which is now fixed it is needed to fix tmp102 platform data.
Swapping the values of slope/offset and slope_cpu/offset_cpu, so that
previous values used for cpu domain are still used by CPU domain.

Change-Id: Ibb0c090a0e3977612857b9e59f67a1a63c870cd1
Signed-off-by: Ruslan Ruslichenko <x0191366@ti.com>