arm: OMAP5: hsmmc: API to Pull up/down SD_CLK
Some UHS1-SD cards are very sensitive to ramp down time
of SD_CLK. Because of this, when the SD_CLK is pulled
up by default, and the PBIAS power is taken down, the
SD_CLK line takes a little time to ramp down. Some SD
cards do not respond beyond this time.

So, provide API to turn Pull up/down the SD_CLK that
can be used to first turn the Pull down on SD_CLK before
the power is turned down on the PBIAS and turn pull up
just before sending the free running SD_CLK.

Change-Id: Idab08488341a73edbfb11bc447252fd06a5f703b
Signed-off-by: Viswanath Puttagunta <vishp@ti.com>