watchdog: omap_wdt: fix runtime PM API usage
The WDTIMER2 HW module will stay in TRANSITION state If the counting is
not stopped and IdleRequest is asserted, in order to go in IDLE state.
Now WDTIMER2 HW module constantly stays in "in transition" state,
because runtime PM API is used wrongly, for example when wdt enabled:
 - pm_runtime_get_sync()
 - enable counting
 - pm_runtime_put_sync()
   <<-- at this point module state is switched to "disabled", but
counter is not stopped, as result WDTIMER2 will stuck "in transition"
state

More over, while WDTIMER2 module is "in transition" state it prevents
system (core) from entering deeper LP states (and on OMAP4 SYSCLK can't
be gated).

The right sequences to work with WDTIMER2 are:
   enable:
   - pm_runtime_get_sync()
   - enable counting

   disable:
   - disable counting
   - pm_runtime_put_sync()

   suspend:
   - do disable()

   resume:
   - do enable()

and pm_runtime_put_sync() shouldn't be called for it while counting is
enabled.

Hence, fix it by using of runtime PM in a right way inside omap_wdt
driver

Change-Id: Iaf40371b4ef2e121c042d73dc13b09a5d3b92072
Signed-off-by: Andrii Tseglytskyi <andrii.tseglytskyi@ti.com>
[ivan.khoronzhuk@ti.com: ported from p-android-omap-3.0]
Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@ti.com>
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>