ARM: OMAP: BOARD ID: handle forced_rev sanely
when boards pass forced_rev, it implies the data is already
hexadecimal format. current algo assumes that if predefined
board matches fail, the "dec" has to be converted back to
hex. In the case of forced_rev, this assumption is wrong
and results in messed up board rev info.

We do the convertion upahead based on forced_rev and then
follow the usual fall throughs.

Change-Id: I8eb06bb5bd2d655231f626be9b3302fe14c69fa9
Signed-off-by: Nishanth Menon <nm@ti.com>