OMAP4: Tablet/SDP4430: Enable McBSP1 DR pull-down
McBSP1 DR pin is used for BT RX data coming from WiLink chip, this
link is configured in I2S mode. BT data is mono and is transferred
on the left channel, the right channel is expected to have 0's.

However, after the last bit of left channel is transferred, DR line
experiments a slow ramp down instead of being driven to low state
as would be expected since the right channel is 0.

The duration of the ramp can impact how McBSP1 samples the first
bit of right channel, which if too long can cause McBSP1 to interpret
the right channel as a negative full-scale value. The higher the
sample rate, the more severe this problem is. For WB (16kHz), this
problem was causing noise in right channel.

McBSP1 DR pull-down has been enabled to reduce the duration of the
ramp, the following values were measured on sEVM.

Sample Freq  Ramp time (no PD)  Ramp time (with PD)
  8kHz            5.88us             1.08us
 16kHz            8.80us             2.92us

Change-Id: I37e87727b0b9b71a4256f83b679ea55424c49605
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>