ARM4: OMAP4+: HDMI: Relocate devices for audio codec and card
Relocate the creation the platform devices for audio the HDMI audio codec and
the audio card to display.c. This allows the display code to create the required
platform devices based on what is wired on the board. Thus, as many devices as
required are created; or none if the HDMI ouptut is not implemented.

Signed-off-by: Ricardo Neri <rneri@dextratech.com>

Conflicts:

	arch/arm/mach-omap2/board-4430sdp.c
	arch/arm/mach-omap2/board-omap4panda.c

Signed-off-by: Archit Taneja <archit@ti.com>