ARM: OMAP4+: HDMI: Relocate the device for audio interface
The HDMI display and audio functionality share the same resources (e.g., register
spaces). The ASoC HDMI CPU-DAI driver needs access to some (but not all) the
resources of the dss_hdmi platform device. As such resources are within the
address space of omapdss_hdmi, it makes sense to have the DSS HDMI driver to
create the platform driver with the required resources.

Signed-off-by: Ricardo Neri <rneri@dextratech.com>

Conflicts:

	arch/arm/mach-omap2/devices.c

Signed-off-by: Archit Taneja <archit@ti.com>