android_omap_defconfig: enable Linux Bluetooth subsystem
For now we are using Bluedroid stack which can work directly with TI shared
transport tty. This is the reason that modules bluetooth.ko and hidp.ko and
other BlueZs modules were excluded from AFS. Bluedroid stack is used to just
enable/disable bluetooth hardware module and runs the hciattach. But actual
bluetooth profile and implementation of bluetooth specification is still
available in BlueZ stack. HIDP layer of the BlueZ stack is needed to use
the bluetooth keyboard as HID input device. So compile BlueZ modules as part
of the kernel.

Change-Id: Ic1979abac46166eb4744deee63bb210a655e7c77
Signed-off-by: Oleksandr Dmytryshyn <oleksandr.dmytryshyn@ti.com>