usb: musb: Switch musb interface for charger detection
To perform charger detection sequence musb interface
should be switched to ULPI_12PIN.
The musb is in suspend before charger detection and
resumes after the detection. Therefore we switch the
interface in suspend. On resume the interface is
restored to it's previous state. Thus switching
to ULPI_12PIN in suspend does not affect it's
active work.

Change-Id: I1d0a1f9fd6c003bf6d398f049fbc60ad37e36478
Signed-off-by: Oleg_Kosheliev <oleg.kosheliev@ti.com>