remoteproc: add support for multiple vdevs within a rproc instance
Support has been added to the core remoteproc code to allow multiple
virtio devices per rproc instance. This feature is generic in general,
but is being added primarily to support booting non-smp ipu cores in
OMAP. The feature is designed by processing all the virtio device
resources first, and registering all of them together. The processor
is booted only when the last virtio device is probed.

Change-Id: I9cbd8f4fad6aeadad0878bab15e2fbba8e154e81
Signed-off-by: Suman Anna <s-anna@ti.com>