rpmsg: do not kick remote processor upon priming rx buffers
The rpmsg bus driver currently sends a notification to the remote
processor after adding all the receive buffers to the vring. This
is commented out to help support the boot of a non-SMP IPU image.

The non-SMP IPU implementation uses two virtio devices, one per
processor, but represented by a single rproc instance. The rpmsg
probe function would cause the processor to be started for each
virtio device, and then notify the remote processor using the
mailbox (which is created only when the rproc is started) within
the rproc instance. This logic will be adjusted for non-SMP IPU,
wherein the processor would be started only when the second virtio
device is instantiated, and hence the notification cannot be made
prior to starting of the processor.

This change is a quick solution (not so neat) and works for the
most part since the first transmission from the remote processor
is always to announce a rpmsg channel, and the tx packet retrieval
for this announcement will be successful only after the rx buffers
are primed by host. The current notification is used for posting a
semaphore, signalling any tx threads that a new buffer is available.

There is still a small chance wherein a tx buffer retrieval may
happen before the rx buffers are primed, but the image initialization
on the remote processor is long enough to minimize these chances
greatly.

Change-Id: I6f5a4a44e50f3084a102bdeb3d2d235c86cde09a
Signed-off-by: Suman Anna <s-anna@ti.com>