omap: remoteproc: add support to boot OMAP4 IPU in non-smp mode
IPU would require an additional timer to provide the system tick
for core1 when running in non-SMP mode. GPTimer 4 is added to the
timers data to serve as this additional timer.

The current logic is hardcoded to boot OMAP4 IPU in non-SMP mode,
with OMAP5 IPU unchanged to boot in SMP-mode. The choice was made
at run-time instead of configuration time, since ARCH_OMAP4 and
ARCH_OMAP5 would both be enabled in most common defconfigs like
omap2plus_defconfig.

Change-Id: Iba94b98e7d1d956f193f7394d797a5e08f416c6b
Signed-off-by: Suman Anna <s-anna@ti.com>