ARM: OMAP4: twl-common: Enable regulator VANA in suspend
The VANA voltage regulator is dedicated to supply
the analog functions of the TWL603x device, such as
the GPADC, fuel gauge, and other analog circuitries.

This patch enables VANA in suspend to provide
power for the fuel gauge. The fuel gauge should
continue measuring of the battery charge/discharge
in suspend. If VANA is off the fuel gauge gets
noise and indicates an incorrect state of charge.
(E.g. when VANA is off in suspend the fuel gauge
could see that the current is about 2000mA of
charging instead of 3mA of discharging which leads to
indicating what the battery was charged without the
power source.)
Measured additional power consumption in suspend is
about 0.07mA max (it is VANA itself, fuel gauge circuit
and other analog consumers).

Change-Id: Ib1be2055a194e3b55c6fbaa52fefeab84e5b9673
Signed-off-by: Oleg_Kosheliev <oleg.kosheliev@ti.com>