ARM: OMAP2+: add remoteproc devices for OMAP4
OMAP4 has two remote processor subsystems - a dual ARM Cortex-M3 IPU
and a TI C6400-based DSP subsystem. This patch adds the support for
creating remoteproc platform devices, and other necessary platform
specific hooks needed for enabling and booting the remote processors.

A OMAP_REMOTEPROC dependent config option is defined for each rproc
device that provides the choice to enable/disable creating the
corresponding rproc device. The clock and reset management for these
devices is provided by the omap_device/omap_hwmod layer and hence
has to be implemented through platform data functions ops. The
omap_device layer does not program the resets when enabling a device
and is handled by the remoteproc platform hooks. The hooks perform
the reset management based on the remoteproc subsystem type. An IPU
subsystem has two resets, and the remoteproc device management code
asserts/deasserts these together because of the intrinsic device
management dependencies such as a shared MMU and shared cache between
the two Cortex-M3 processors. This allows the OMAP IPU to be run in
either a SMP-mode or a non-SMP mode (using a post-processed combined
image), and is transparent to the remoteproc core driver code.

Each of the devices is bound to its dedicated IOMMU device, used for
managing the MMU; and a dedicated mailbox instance for signalling
message availability and inter processor communication. These slave
devices are used by the remoteproc driver core through the respective
API provided by the iommu and mailbox drivers. A dedicated CMA pool,
tied with each remoteproc device, is used for managing the various
memory needs of the remote processors.

The current CMA pools and sizes are defined statically for each
device. The starting addresses are fixed to meet current dependencies
on the remote processor side, and will go away when the remote-side
code has been improved to gather this information runtime during its
initialization. There are no Kconfig symbols associated with the CMA
bases and sizes, since it would not scale well for remoteproc devices
on multiple SoCs without making the configuration code cumbersome.

The remoteproc CMA pools are also defined to be within the first 512MB
of RAM and end exactly at 0xA000000. The CMA base addresses are
computed by subtracting the required size from 0xA0000000. This allows
a vmalloc size of upto 496MB [__pa(VMALLOC_START) value of 0xA0800000],
without eliciting invalid pfn warnings when the cma regions are
activated.

Based on patches developed previously for TI product kernels, and
includes contributions from:
    Fernando Guzman Lugo <fernando.lugo@ti.com>
    Ohad Ben-Cohen <ohad@wizery.com>
    Juan Gutierrez <jgutierrez@ti.com>

Signed-off-by: Ohad Ben-Cohen <ohad@wizery.com>
Signed-off-by: Suman Anna <s-anna@ti.com>