ARM: OMAP2+: pass a SoC identifier to omap_rproc_reserve_cma
OMAP5 has to support a larger memory footprint compared to
OMAP4. Support has been added to be able to pick different
CMA addresses and sizes based on a SoC identifier for that
SoC. These SoC identifiers are required since the SoC/CPU
revision check api is not available until the machine
init_early, and the CMA pool reservation for the remoteproc
devices has to be performed earlier during the board memory
setup phase.

Signed-off-by: Suman Anna <s-anna@ti.com>