PM / AVS: SmartReflex: remove sr_late_init function
sr_late_init() is actually not needed and, more over, contains wrong
errors handling code. So, it can be simply removed by adding
corresponding actions in omap_sr_probe().

As part of this patch the SR auto-enable is moved to the end of
and static input parameters check on top of omap_sr_probe().

Change-Id: I1784a1547d1848aa727af3cccd4de4342cfe383d
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>