ARM: OMAP2+: AM33XX: add support for mailbox used for WKUP_M3
Mailbox IP on AM33XX is the same as that present in OMAP4.
The single instance of Mailbox IP on AM33XX contains
8 sub-modules and facilitates communication between MPU,
PRUs and WKUP_M3.

The first mailbox sub-module is assigned for communication
between MPU and WKUP-M3. The mailbox is used only for
interrupting WKUP-M3, and MPU is responsible for clearing
the messages and the mailbox IP interrupt for WKUP-M3.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>
[s-anna@ti.com: minor updates to commit msg & variables]
Signed-off-by: Suman Anna <s-anna@ti.com>