OMAPDSS: DSS: corrected disabling workaround of WB.
This change fixes the disconnecting of pipes from a manager,
which was connected to WB in m2m mode. Now WB workaround
disconnects all the pipes which was connected to WB.

Was the usecase when vid2 and vid3 was connected to WB,
next frame was with vid2 only (vid3 - off) and 3rd frame
was request to disable WB. WB workaround was called only
for vid2. When HWC switches vid3 to another manager -
SyncLost issue was happening.

Change-Id: I0e95e173d3406f88c1c6b65ff6aabbfa1c113e74
Signed-off-by: Artem Rudenko <artemrudenko@ti.com>