OMAPDSS: DSSCOMP: do not clone from disabled layer
In case of cloning layer from other overlay, we should always check
whether overlay we want to clone from is valid, or we may run
into severe troubles. One example of failing case is when 1D mapping
of overlay fails, we can't clone from it, as it will be disabled and
we will eventually end up with memory violation.

Change-Id: Ic1eac90119fc4c08bad783c767ba499286071b22
Signed-off-by: Sergiy Kibrik <sergiy.kibrik@globallogic.com>